+++ src/main/java/insane96mcp/iguanatweaksreborn/data/generator/client/ITRItemModelsProvider.java
@@ -2,7 +2,7 @@ package insane96mcp.iguanatweaksreborn.data.generator.client;
 
 import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.farming.crops.Crops;
-import insane96mcp.iguanatweaksreborn.module.world.BerryBushes;
+import insane96mcp.iguanatweaksreborn.module.world.Berries;
 import insane96mcp.iguanatweaksreborn.module.world.CyanFlower;
 import net.minecraft.data.PackOutput;
 import net.minecraft.resources.ResourceLocation;
@@ -27,7 +27,7 @@ public class ITRItemModelsProvider extends ItemModelProvider {
         basicItem(Crops.ROOTED_ONION.get());
         basicItem(Crops.RICE_SEEDS.get());
 
-        basicItem(BerryBushes.SWEET_BERRY_SEEDS.get());
+        basicItem(Berries.SWEET_BERRY_SEEDS.get());
 
         basicItemWithTexture(CyanFlower.FLOWER.item().get(), new ResourceLocation(IguanaTweaksReborn.MOD_ID, "block/cyan_flower"));
         basicItemWithTexture(Crops.SOLANUM_NEOROSSII.item().get(), new ResourceLocation(IguanaTweaksReborn.MOD_ID, "block/solanum_neorossii"));